+++ kikuchipy/indexing/__init__.py
@@ -19,8 +19,8 @@
 """Tools for indexing of EBSD patterns by comparison to simulated
 patterns.
 
-See the EBSD method :meth:`~kikuchipy.signals.EBSD.dictionary_indexing`
-for easy dictionary indexing, which uses these tools.
+The EBSD method :meth:`~kikuchipy.signals.EBSD.match_patterns` uses
+these tools for pattern matching.
 """
 
 from kikuchipy.indexing._merge_crystal_maps import merge_crystal_maps
@@ -28,13 +28,11 @@ from kikuchipy.indexing.orientation_similarity_map import (
     orientation_similarity_map,
 )
 from kikuchipy.indexing import similarity_metrics
-from kikuchipy.indexing._static_dictionary_indexing import (
-    StaticDictionaryIndexing,
-)
+from kikuchipy.indexing._static_pattern_matching import StaticPatternMatching
 
 __all__ = [
     "merge_crystal_maps",
     "orientation_similarity_map",
     "similarity_metrics",
-    "StaticDictionaryIndexing",
+    "StaticPatternMatching",
 ]