+++ kikuchipy/indexing/similarity_metrics.py
@@ -50,8 +50,8 @@ def make_similarity_metric(
     equal size.
 
     This factory function wraps metric functions for use in
-    :meth:`~kikuchipy.signals.EBSD.dictionary_indexing`, which uses
-    :class:`~kikuchipy.indexing.StaticDictionaryIndexing`.
+    :meth:`~kikuchipy.signals.EBSD.match_patterns` (which uses
+    :class:`~kikuchipy.indexing.StaticPatternMatching`).
 
     Parameters
     ----------