+++ kikuchipy/signals/ebsd.py
@@ -51,7 +51,7 @@ from kikuchipy.pattern._pattern import (
     fft_filter,
     _dynamic_background_frequency_space_setup,
 )
-from kikuchipy.indexing import StaticDictionaryIndexing
+from kikuchipy.indexing import StaticPatternMatching
 from kikuchipy.indexing.similarity_metrics import SimilarityMetric
 from kikuchipy.signals.util._metadata import (
     ebsd_metadata,
@@ -195,7 +195,7 @@ class EBSD(CommonImage, Signal2D):
 
         See Also
         --------
-        kikuchipy.signals.EBSD.set_phase_parameters
+        ~kikuchipy.signals.EBSD.set_phase_parameters
 
         Examples
         --------
@@ -297,7 +297,7 @@ class EBSD(CommonImage, Signal2D):
 
         See Also
         --------
-        kikuchipy.signals.EBSD.set_experimental_parameters
+        ~kikuchipy.signals.EBSD.set_experimental_parameters
 
         Examples
         --------
@@ -356,7 +356,7 @@ class EBSD(CommonImage, Signal2D):
 
         See Also
         --------
-        kikuchipy.signals.EBSD.set_detector_calibration
+        ~kikuchipy.signals.EBSD.set_detector_calibration
 
         Examples
         --------
@@ -382,7 +382,7 @@ class EBSD(CommonImage, Signal2D):
 
         See Also
         --------
-        kikuchipy.signals.EBSD.set_scan_calibration
+        ~kikuchipy.signals.EBSD.set_scan_calibration
 
         Examples
         --------
@@ -430,7 +430,7 @@ class EBSD(CommonImage, Signal2D):
 
         See Also
         --------
-        kikuchipy.signals.EBSD.remove_dynamic_background,
+        ~kikuchipy.signals.EBSD.remove_dynamic_background
 
         Examples
         --------
@@ -566,9 +566,9 @@ class EBSD(CommonImage, Signal2D):
 
         See Also
         --------
-        kikuchipy.signals.EBSD.remove_static_background,
-        kikuchipy.signals.EBSD.get_dynamic_background,
-        kikuchipy.pattern.remove_dynamic_background,
+        kikuchipy.signals.EBSD.remove_static_background
+        kikuchipy.signals.EBSD.get_dynamic_background
+        kikuchipy.pattern.remove_dynamic_background
         kikuchipy.pattern.get_dynamic_background
 
         Examples
@@ -757,8 +757,8 @@ class EBSD(CommonImage, Signal2D):
 
         See also
         --------
-        kikuchipy.signals.EBSD.rescale_intensity,
-        kikuchipy.signals.EBSD.normalize_intensity
+        ~kikuchipy.signals.EBSD.rescale_intensity
+        ~kikuchipy.signals.EBSD.normalize_intensity
 
         Examples
         --------
@@ -880,7 +880,7 @@ class EBSD(CommonImage, Signal2D):
 
         return image_quality_map
 
-    def dictionary_indexing(
+    def match_patterns(
         self,
         simulations,
         metric: Union[str, SimilarityMetric] = "ncc",
@@ -889,10 +889,16 @@ class EBSD(CommonImage, Signal2D):
         return_merged_crystal_map: bool = False,
         get_orientation_similarity_map: bool = False,
     ) -> Union[CrystalMap, List[CrystalMap]]:
-        """Perform dictionary indexing :cite:`chen2015dictionary` by
-        comparing experimental patterns against pre-computed simulations
-        with normalized cross-correlation or a user-defined similarity
-        metric.
+        """Match each experimental pattern to all simulated patterns, of
+        known crystal orientations in pre-computed dictionaries
+        :cite:`chen2015dictionary,jackson2019dictionary`, to determine
+        their phase and orientation.
+
+        A suitable similarity metric, the normalized cross-correlation
+        (:func:`~kikuchipy.indexing.similarity_metrics.ncc`), is used by
+        default, but a valid user-defined similarity metric may be used
+        instead (see
+        :func:`~kikuchipy.indexing.similarity_metrics.make_similarity_metric`).
 
         :class:`~orix.crystal_map.crystal_map.CrystalMap`'s for each
         dictionary with "scores" and "simulation_indices" as properties
@@ -936,8 +942,13 @@ class EBSD(CommonImage, Signal2D):
         :func:`~kikuchipy.indexing.merge_crystal_maps` and
         :func:`~kikuchipy.indexing.orientation_similarity_map`,
         respectively.
+
+        See Also
+        --------
+        ~kikuchipy.indexing.similarity_metrics.make_similarity_metric
+        ~kikuchipy.indexing.similarity_metrics.ndp
         """
-        sdi = StaticDictionaryIndexing(simulations)
+        sdi = StaticPatternMatching(simulations)
         return sdi(
             signal=self,
             metric=metric,
@@ -997,7 +1008,7 @@ class EBSD(CommonImage, Signal2D):
 
         See Also
         --------
-        :class:`~kikuchipy.filters.window.Window`
+        ~kikuchipy.filters.window.Window
         """
         dtype_out = self.data.dtype
 
@@ -1086,8 +1097,8 @@ class EBSD(CommonImage, Signal2D):
 
         See Also
         --------
-        :class:`~kikuchipy.filters.window.Window`,
-        :func:`scipy.signal.windows.get_window`,
+        ~kikuchipy.filters.window.Window
+        :func:`scipy.signal.windows.get_window`
         :func:`scipy.ndimage.correlate`
 
         Examples
@@ -1261,7 +1272,7 @@ class EBSD(CommonImage, Signal2D):
 
         See Also
         --------
-        kikuchipy.signals.EBSD.get_virtual_bse_intensity
+        ~kikuchipy.signals.EBSD.get_virtual_bse_intensity
         """
         # Plot signal if necessary
         if self._plot is None or not self._plot.is_active:
@@ -1339,7 +1350,7 @@ class EBSD(CommonImage, Signal2D):
 
         See Also
         --------
-        kikuchipy.signals.EBSD.plot_virtual_bse_intensity
+        ~kikuchipy.signals.EBSD.plot_virtual_bse_intensity
         """
         vbse = roi(self, axes=self.axes_manager.signal_axes)
         vbse_sum = self._get_sum_signal(vbse, out_signal_axes)
@@ -1391,7 +1402,7 @@ class EBSD(CommonImage, Signal2D):
 
         See Also
         --------
-        kikuchipy.io.plugins.h5ebsd.file_writer,\
+        kikuchipy.io.plugins.h5ebsd.file_writer
         kikuchipy.io.plugins.nordif.file_writer
         """
         if filename is None: